# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DistributorOS API",
    "description": "DistributorOS distribution management API.",
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/v1/business/settings": {
      "get": {
        "tags": [
          "Business Settings"
        ],
        "summary": "Get Business Settings",
        "operationId": "get_business_settings_api_v1_business_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Business Settings"
        ],
        "summary": "Update Business Settings",
        "operationId": "update_business_settings_api_v1_business_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BusinessSettingsResponse": {
        "properties": {
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "currency": {
            "type": "string",
            "enum": [
              "INR",
              "USD",
              "EUR",
              "GBP",
              "AED",
              "SAR",
              "SGD",
              "MYR"
            ],
            "title": "Currency"
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "ml"
            ],
            "title": "Language"
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "title": "Theme"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "business_name",
          "currency",
          "language",
          "theme",
          "timezone"
        ],
        "title": "BusinessSettingsResponse"
      },
      "BusinessSettingsUpdateRequest": {
        "properties": {
          "business_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Name"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "INR",
                  "USD",
                  "EUR",
                  "GBP",
                  "AED",
                  "SAR",
                  "SGD",
                  "MYR"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "en",
                  "ml"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "light",
                  "dark",
                  "system"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BusinessSettingsUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}